pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-cli"
version = "1.0.0"
description = "A colorful CLI weather app"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests",
    "rich",
    "python-dotenv"
]

[project.scripts]
weather = "weather.cli:main"

[tool.setuptools]
packages = ["weather"]

